+++ tests/testthat/test-qenv_get_code.R
@@ -290,6 +290,20 @@ testthat::describe("get_code for specific names", {
       )
     }
   )
+  testthat::it(
+    "doesn't consider function called on the lhs as a dependent in this call (dependency in further calls)",
+    {
+      code <- c(
+        "object_list <- list(x = iris, y = iris)",
+        "object_list_2 <- list(x = mtcars, y = mtcars)",
+        "object_list_2[c('x')] <- c('string')",
+        "object_list[c('x')] <- c('string')"
+      )
+      q <- eval_code(qenv(), code = code)
+      result <- get_code(q, names = "object_list")
+      testthat::expect_identical(result, paste(code[c(1, 4)], collapse = "\n"))
+    }
+  )
 })
 
 
@@ -581,7 +595,7 @@ testthat::test_that("detects occurrence of a function definition when a formal i
 })
 
 testthat::test_that("detects occurrence of a function definition with a @linksto usage", {
-  code <- c(
+  code <- trimws(c(
     "
         foo <- function() {
           env <- parent.frame()
@@ -589,7 +603,7 @@ testthat::test_that("detects occurrence of a function definition with a @linksto
         }",
     "foo() # @linksto x",
     "y <- x"
-  )
+  ))
   q <- eval_code(qenv(), code)
   testthat::expect_identical(
     get_code(q, names = "x"),
@@ -601,7 +615,7 @@ testthat::test_that("detects occurrence of a function definition with a @linksto
 # for loop --------------------------------------------------------------------------------------------------------
 
 testthat::test_that("objects in for loop are extracted if passed as one character", {
-  code <- "
+  code <- trimws("
     some_other_dataset <- mtcars
     original_dataset <- iris[, 1:4]
     count <- 1
@@ -610,11 +624,11 @@ testthat::test_that("objects in for loop are extracted if passed as one characte
       count <- count + 1
     }
     output <- rlang::list2(x = original_dataset)
-  "
+  ")
   q <- eval_code(qenv(), code)
   testthat::expect_identical(
     get_code(q, names = "output"),
-    gsub("\n    some_other_dataset <- mtcars\n", "", code, fixed = TRUE)
+    gsub("some_other_dataset <- mtcars\n", "", code, fixed = TRUE)
   )
 })